api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "stmball", email = "stmball@gmail.com" }
]
dependencies = [
    "shapely>=2.1.0",
    "matplotlib>=3.10.3",
    "numpy>=2.2.5",
    "scikit-image>=0.25.2",
    "torch>=2.7.0",
    "torchvision>=0.22.0",
    "flask>=3.1.1",
    "pillow>=11.2.1",
    "filesplit>=4.1.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/api"]
